+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Eøs/Valutakurs/ValutakursTabellRad.tsx
@@ -4,11 +4,8 @@ import { Table } from '@navikt/ds-react';
 
 import { PeriodeValutakurs } from './PeriodeValutakurs';
 import { StatusOgBarnValutakurs } from './StatusOgBarnValutakurs';
+import { useValutakursSkjema, valutakursFeilmeldingId } from './useValutakursSkjema';
 import ValutakursTabellRadEndre from './ValutakursTabellRadEndre';
-import {
-    useValutakursSkjema,
-    valutakursFeilmeldingId,
-} from '../../../../../context/Valutakurs/ValutakursSkjemaContext';
 import type { IBehandling } from '../../../../../typer/behandling';
 import type { OptionType } from '../../../../../typer/common';
 import type { IRestValutakurs } from '../../../../../typer/eøsPerioder';
