+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Eøs/Valutakurs/useValutakurs.ts
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react';
 
-import type { IBehandling } from '../../typer/behandling';
-import type { IRestValutakurs } from '../../typer/eøsPerioder';
-import { EøsPeriodeStatus } from '../../typer/eøsPerioder';
-import { sorterEøsPerioder } from '../../utils/eøs';
+import type { IBehandling } from '../../../../../typer/behandling';
+import type { IRestValutakurs } from '../../../../../typer/eøsPerioder';
+import { EøsPeriodeStatus } from '../../../../../typer/eøsPerioder';
+import { sorterEøsPerioder } from '../../../../../utils/eøs';
 
 interface IProps {
     åpenBehandling: IBehandling;
